// APVN.LeadsPlatform.Client/apvn-leadsplatform-client/src/app/models/LeadsActionExtends.ts
export class LeadsActionExtends {
    constructor() {
        this.Id = 0;
        this.Index = 0;
    }
    Id: number;
    LeadsId: number;
    LeadsActionId: number;
    SecondHand: number;
    CityId: number;
    Type: number;
    Source: number;
    CreatedDate: Date;
    LastModifiedDate: Date;
    CreatedBy: string;
    LastModifiedBy: string;
    OtherInformation: string;
    LoanMoney: number;
    LoanTime: number;
    BankId: number;
    BargainCurrentPrices: number;
    BargainPrices: number;
    Index: number;
}